backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.11
alembic==1.13.1
openai==1.3.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
mcp==1.0.0
asyncpg==0.29.0
sse-starlette==1.6.5
cryptography==41.0.8
PyJWT==2.8.0
pyotp==2.9.0
qrcode[pil]==7.4.2